refactor: add schema checks to migrations to prevent duplicate column creation or dropping

+++ database/migrations/2026_05_07_033021_add_missing_stat_column_to_pallet_delivery_stats.php
@@ -1,35 +1,34 @@
 <?php
+/*
+ * Author: Raul Perusquia <raul@inikoo.com>
+ * Created: Thu, 07 May 2026 16:42:05 Malaysia Time, Kuala Lumpur, Malaysia
+ * Copyright (c) 2026, Raul A Perusquia Flores
+ */
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
-    {
-        Schema::table('pallet_delivery_stats', function (Blueprint $table) {            
-            $table->integer('number_pallets_state_not_picked')->default(0);
+return new class extends Migration {
 
+    public function up(): void
+    {
+        Schema::table('pallet_delivery_stats', function (Blueprint $table) {
+            if (!Schema::hasColumn('pallet_delivery_stats', 'number_pallets_state_not_picked')) {
+                $table->integer('number_pallets_state_not_picked')->default(0);
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
+
+    public function down(): void
     {
         Schema::table('pallet_delivery_stats', function (Blueprint $table) {
-            $table->dropColumn([
-                'number_pallets_state_not_picked'
-            ]);
+            if (Schema::hasColumn('pallet_delivery_stats', 'number_pallets_state_not_picked')) {
+                $table->dropColumn([
+                    'number_pallets_state_not_picked'
+                ]);
+            }
         });
     }
 };
